+++ config.c
@@ -1544,6 +1544,7 @@ static void
 free_key_binding(struct config_key_binding *binding)
 {
     free_binding_aux(&binding->aux);
+    tll_free_and_free(binding->modifiers, free);
 }
 
 static void NOINLINE
@@ -1559,43 +1560,26 @@ free_key_binding_list(struct config_key_binding_list *bindings)
     bindings->count = 0;
 }
 
-static bool NOINLINE
-parse_modifiers(struct context *ctx, const char *text, size_t len,
-                struct config_key_modifiers *modifiers)
+static void NOINLINE
+parse_modifiers(const char *text, size_t len, config_modifier_list_t *modifiers)
 {
-    bool ret = false;
-
-    *modifiers = (struct config_key_modifiers){0};
+    tll_free_and_free(*modifiers, free);
 
     /* Handle "none" separately because e.g. none+shift is nonsense */
     if (strncmp(text, "none", len) == 0)
-        return true;
+        return;
 
     char *copy = xstrndup(text, len);
 
-    for (char *tok_ctx = NULL, *key = strtok_r(copy, "+", &tok_ctx);
+    for (char *ctx = NULL, *key = strtok_r(copy, "+", &ctx);
          key != NULL;
-         key = strtok_r(NULL, "+", &tok_ctx))
+         key = strtok_r(NULL, "+", &ctx))
     {
-        if (streq(key, XKB_MOD_NAME_SHIFT))
-            modifiers->shift = true;
-        else if (streq(key, XKB_MOD_NAME_CTRL))
-            modifiers->ctrl = true;
-        else if (streq(key, XKB_MOD_NAME_ALT))
-            modifiers->alt = true;
-        else if (streq(key, XKB_MOD_NAME_LOGO))
-            modifiers->super = true;
-        else {
-            LOG_CONTEXTUAL_ERR("not a valid modifier name: %s", key);
-            goto out;
-        }
+        tll_push_back(*modifiers, xstrdup(key));
     }
 
-    ret = true;
-
-out:
     free(copy);
-    return ret;
+    tll_sort(*modifiers, strcmp);
 }
 
 static int NOINLINE
@@ -1731,6 +1715,7 @@ value_to_key_combos(struct context *ctx, int action,
 
     /* Count number of combinations */
     size_t combo_count = 1;
+    size_t used_combos = 0;  /* For error handling */
     for (const char *p = strchr(ctx->value, ' ');
          p != NULL;
          p = strchr(p + 1, ' '))
@@ -1746,7 +1731,7 @@ value_to_key_combos(struct context *ctx, int action,
     for (char *tok_ctx = NULL, *combo = strtok_r(copy, " ", &tok_ctx);
          combo != NULL;
          combo = strtok_r(NULL, " ", &tok_ctx),
-             idx++)
+             idx++, used_combos++)
     {
         struct config_key_binding *new_combo = &new_combos[idx];
         new_combo->action = action;
@@ -1757,6 +1742,7 @@ value_to_key_combos(struct context *ctx, int action,
         new_combo->aux.master_copy = idx == 0;
         new_combo->aux.pipe = *argv;
 #endif
+        memset(&new_combo->modifiers, 0, sizeof(new_combo->modifiers));
         new_combo->path = ctx->path;
         new_combo->lineno = ctx->lineno;
 
@@ -1765,11 +1751,9 @@ value_to_key_combos(struct context *ctx, int action,
         if (key == NULL) {
             /* No modifiers */
             key = combo;
-            new_combo->modifiers = (struct config_key_modifiers){0};
         } else {
             *key = '\0';
-            if (!parse_modifiers(ctx, combo, key - combo, &new_combo->modifiers))
-                goto err;
+            parse_modifiers(combo, key - combo, &new_combo->modifiers);
             key++;  /* Skip past the '+' */
         }
 
@@ -1779,6 +1763,7 @@ value_to_key_combos(struct context *ctx, int action,
             new_combo->k.sym = xkb_keysym_from_name(key, 0);
             if (new_combo->k.sym == XKB_KEY_NoSymbol) {
                 LOG_CONTEXTUAL_ERR("not a valid XKB key name: %s", key);
+                free_key_binding(new_combo);
                 goto err;
             }
             break;
@@ -1799,6 +1784,7 @@ value_to_key_combos(struct context *ctx, int action,
                         LOG_CONTEXTUAL_ERRNO("invalid click count: %s", _count);
                     else
                         LOG_CONTEXTUAL_ERR("invalid click count: %s", _count);
+                    free_key_binding(new_combo);
                     goto err;
                 }
 
@@ -1808,6 +1794,7 @@ value_to_key_combos(struct context *ctx, int action,
             new_combo->m.button = mouse_button_name_to_code(key);
             if (new_combo->m.button < 0) {
                 LOG_CONTEXTUAL_ERR("invalid mouse button name: %s", key);
+                free_key_binding(new_combo);
                 goto err;
             }
 
@@ -1838,41 +1825,89 @@ value_to_key_combos(struct context *ctx, int action,
     return true;
 
 err:
+    for (size_t i = 0; i < used_combos; i++)
+        free_key_binding(&new_combos[i]);
     free(copy);
     return false;
 }
 
 static bool
-modifiers_equal(const struct config_key_modifiers *mods1,
-                const struct config_key_modifiers *mods2)
+modifiers_equal(const config_modifier_list_t *mods1,
+                const config_modifier_list_t *mods2)
 {
-    bool shift = mods1->shift == mods2->shift;
-    bool alt = mods1->alt == mods2->alt;
-    bool ctrl = mods1->ctrl == mods2->ctrl;
-    bool super = mods1->super == mods2->super;
-    return shift && alt && ctrl && super;
+    if (tll_length(*mods1) != tll_length(*mods2))
+        return false;
+
+    size_t count = 0;
+    tll_foreach(*mods1, it1) {
+        size_t skip = count;
+        tll_foreach(*mods2, it2) {
+            if (skip > 0) {
+                skip--;
+                continue;
+            }
+
+            if (strcmp(it1->item, it2->item) != 0)
+                return false;
+            break;
+        }
+
+        count++;
+    }
+
+    return true;
+    /*
+     * bool shift = mods1->shift == mods2->shift;
+     * bool alt = mods1->alt == mods2->alt;
+     * bool ctrl = mods1->ctrl == mods2->ctrl;
+     * bool super = mods1->super == mods2->super;
+     * return shift && alt && ctrl && super;
+     */
+}
+
+UNITTEST
+{
+    config_modifier_list_t mods1 = tll_init();
+    config_modifier_list_t mods2 = tll_init();
+
+    tll_push_back(mods1, xstrdup("foo"));
+    tll_push_back(mods1, xstrdup("bar"));
+
+    tll_push_back(mods2, xstrdup("foo"));
+    xassert(!modifiers_equal(&mods1, &mods2));
+
+    tll_push_back(mods2, xstrdup("zoo"));
+    xassert(!modifiers_equal(&mods1, &mods2));
+
+    free(tll_pop_back(mods2));
+    tll_push_back(mods2, xstrdup("bar"));
+    xassert(modifiers_equal(&mods1, &mods2));
+
+    tll_free_and_free(mods1, free);
+    tll_free_and_free(mods2, free);
 }
 
 static bool
-modifiers_disjoint(const struct config_key_modifiers *mods1,
-                const struct config_key_modifiers *mods2)
+modifiers_disjoint(const config_modifier_list_t *mods1,
+                const config_modifier_list_t *mods2)
 {
-    bool shift = mods1->shift && mods2->shift;
-    bool alt = mods1->alt && mods2->alt;
-    bool ctrl = mods1->ctrl && mods2->ctrl;
-    bool super = mods1->super && mods2->super;
-    return !(shift || alt || ctrl || super);
+    return !modifiers_equal(mods1, mods2);
 }
 
 static char * NOINLINE
-modifiers_to_str(const struct config_key_modifiers *mods)
+modifiers_to_str(const config_modifier_list_t *mods)
 {
-    char *ret = xasprintf(
-        "%s%s%s%s",
-        mods->ctrl ? XKB_MOD_NAME_CTRL "+" : "",
-        mods->alt ? XKB_MOD_NAME_ALT "+": "",
-        mods->super ? XKB_MOD_NAME_LOGO "+": "",
-        mods->shift ? XKB_MOD_NAME_SHIFT "+": "");
+    size_t len = tll_length(*mods);  /* ‘+’ , and NULL terminator */
+    tll_foreach(*mods, it)
+        len += strlen(it->item);
+
+    char *ret = xmalloc(len);
+    size_t idx = 0;
+    tll_foreach(*mods, it) {
+        idx += snprintf(&ret[idx], len - idx, "%s", it->item);
+        ret[idx++] = '+';
+    }
+    ret[--idx] = '\0';
     return ret;
 }
 
@@ -2014,10 +2049,13 @@ UNITTEST
     xassert(bindings.arr[0].action == TEST_ACTION_FOO);
     xassert(bindings.arr[1].action == TEST_ACTION_BAR);
     xassert(bindings.arr[1].k.sym == XKB_KEY_g);
-    xassert(bindings.arr[1].modifiers.ctrl);
+    xassert(tll_length(bindings.arr[1].modifiers) == 1);
+    xassert(strcmp(tll_front(bindings.arr[1].modifiers), XKB_MOD_NAME_CTRL) == 0);
     xassert(bindings.arr[2].action == TEST_ACTION_BAR);
     xassert(bindings.arr[2].k.sym == XKB_KEY_x);
-    xassert(bindings.arr[2].modifiers.ctrl && bindings.arr[2].modifiers.shift);
+    xassert(tll_length(bindings.arr[2].modifiers) == 2);
+    xassert(strcmp(tll_front(bindings.arr[2].modifiers), XKB_MOD_NAME_CTRL) == 0);
+    xassert(strcmp(tll_back(bindings.arr[2].modifiers), XKB_MOD_NAME_SHIFT) == 0);
 
     /*
      * REPLACE foo with foo=Mod+v Shift+q
@@ -2033,10 +2071,12 @@ UNITTEST
     xassert(bindings.arr[1].action == TEST_ACTION_BAR);
     xassert(bindings.arr[2].action == TEST_ACTION_FOO);
     xassert(bindings.arr[2].k.sym == XKB_KEY_v);
-    xassert(bindings.arr[2].modifiers.alt);
+    xassert(tll_length(bindings.arr[2].modifiers) == 1);
+    xassert(strcmp(tll_front(bindings.arr[2].modifiers), XKB_MOD_NAME_ALT) == 0);
     xassert(bindings.arr[3].action == TEST_ACTION_FOO);
     xassert(bindings.arr[3].k.sym == XKB_KEY_q);
-    xassert(bindings.arr[3].modifiers.shift);
+    xassert(tll_length(bindings.arr[3].modifiers) == 1);
+    xassert(strcmp(tll_front(bindings.arr[3].modifiers), XKB_MOD_NAME_SHIFT) == 0);
 
     /*
      * REMOVE bar
@@ -2103,7 +2143,7 @@ resolve_key_binding_collisions(struct config *conf, const char *section_name,
         struct config_key_binding *binding1 = &bindings->arr[i];
         xassert(binding1->action != BIND_ACTION_NONE);
 
-        const struct config_key_modifiers *mods1 = &binding1->modifiers;
+        const config_modifier_list_t *mods1 = &binding1->modifiers;
 
         /* Does our modifiers collide with the selection override mods? */
         if (type == MOUSE_BINDING &&
@@ -2127,7 +2167,7 @@ resolve_key_binding_collisions(struct config *conf, const char *section_name,
                 continue;
             }
 
-            const struct config_key_modifiers *mods2 = &binding2->modifiers;
+            const config_modifier_list_t *mods2 = &binding2->modifiers;
 
             bool mods_equal = modifiers_equal(mods1, mods2);
             bool sym_equal;
@@ -2252,13 +2292,9 @@ parse_section_mouse_bindings(struct context *ctx)
     const char *value = ctx->value;
 
     if (streq(key, "selection-override-modifiers")) {
-        if (!parse_modifiers(
-                ctx, ctx->value, strlen(value),
-                &conf->mouse.selection_override_modifiers))
-        {
-            LOG_CONTEXTUAL_ERR("%s: invalid modifiers '%s'", key, ctx->value);
-            return false;
-        }
+        parse_modifiers(
+            ctx->value, strlen(value),
+            &conf->mouse.selection_override_modifiers);
         return true;
     }
 
@@ -2830,37 +2866,38 @@ get_server_socket_path(void)
     return xasprintf("%s/foot-%s.sock", xdg_runtime, wayland_display);
 }
 
-#define m_none           {0}
-#define m_alt            {.alt = true}
-#define m_ctrl           {.ctrl = true}
-#define m_shift          {.shift = true}
-#define m_ctrl_shift     {.ctrl = true, .shift = true}
-#define m_ctrl_shift_alt {.ctrl = true, .shift = true, .alt = true}
+static config_modifier_list_t
+m(const char *text)
+{
+    config_modifier_list_t ret = tll_init();
+    parse_modifiers(text, strlen(text), &ret);
+    return ret;
+}
 
 static void
 add_default_key_bindings(struct config *conf)
 {
-    static const struct config_key_binding bindings[] = {
-        {BIND_ACTION_SCROLLBACK_UP_PAGE, m_shift, {{XKB_KEY_Prior}}},
-        {BIND_ACTION_SCROLLBACK_DOWN_PAGE, m_shift, {{XKB_KEY_Next}}},
-        {BIND_ACTION_CLIPBOARD_COPY, m_ctrl_shift, {{XKB_KEY_c}}},
-        {BIND_ACTION_CLIPBOARD_COPY, m_none, {{XKB_KEY_XF86Copy}}},
-        {BIND_ACTION_CLIPBOARD_PASTE, m_ctrl_shift, {{XKB_KEY_v}}},
-        {BIND_ACTION_CLIPBOARD_PASTE, m_none, {{XKB_KEY_XF86Paste}}},
-        {BIND_ACTION_PRIMARY_PASTE, m_shift, {{XKB_KEY_Insert}}},
-        {BIND_ACTION_SEARCH_START, m_ctrl_shift, {{XKB_KEY_r}}},
-        {BIND_ACTION_FONT_SIZE_UP, m_ctrl, {{XKB_KEY_plus}}},
-        {BIND_ACTION_FONT_SIZE_UP, m_ctrl, {{XKB_KEY_equal}}},
-        {BIND_ACTION_FONT_SIZE_UP, m_ctrl, {{XKB_KEY_KP_Add}}},
-        {BIND_ACTION_FONT_SIZE_DOWN, m_ctrl, {{XKB_KEY_minus}}},
-        {BIND_ACTION_FONT_SIZE_DOWN, m_ctrl, {{XKB_KEY_KP_Subtract}}},
-        {BIND_ACTION_FONT_SIZE_RESET, m_ctrl, {{XKB_KEY_0}}},
-        {BIND_ACTION_FONT_SIZE_RESET, m_ctrl, {{XKB_KEY_KP_0}}},
-        {BIND_ACTION_SPAWN_TERMINAL, m_ctrl_shift, {{XKB_KEY_n}}},
-        {BIND_ACTION_SHOW_URLS_LAUNCH, m_ctrl_shift, {{XKB_KEY_o}}},
-        {BIND_ACTION_UNICODE_INPUT, m_ctrl_shift, {{XKB_KEY_u}}},
-        {BIND_ACTION_PROMPT_PREV, m_ctrl_shift, {{XKB_KEY_z}}},
-        {BIND_ACTION_PROMPT_NEXT, m_ctrl_shift, {{XKB_KEY_x}}},
+    const struct config_key_binding bindings[] = {
+        {BIND_ACTION_SCROLLBACK_UP_PAGE, m(XKB_MOD_NAME_SHIFT), {{XKB_KEY_Prior}}},
+        {BIND_ACTION_SCROLLBACK_DOWN_PAGE, m(XKB_MOD_NAME_SHIFT), {{XKB_KEY_Next}}},
+        {BIND_ACTION_CLIPBOARD_COPY, m(XKB_MOD_NAME_CTRL "+" XKB_MOD_NAME_SHIFT), {{XKB_KEY_c}}},
+        {BIND_ACTION_CLIPBOARD_COPY, m("none"), {{XKB_KEY_XF86Copy}}},
+        {BIND_ACTION_CLIPBOARD_PASTE, m(XKB_MOD_NAME_CTRL "+" XKB_MOD_NAME_SHIFT), {{XKB_KEY_v}}},
+        {BIND_ACTION_CLIPBOARD_PASTE, m("none"), {{XKB_KEY_XF86Paste}}},
+        {BIND_ACTION_PRIMARY_PASTE, m(XKB_MOD_NAME_SHIFT), {{XKB_KEY_Insert}}},
+        {BIND_ACTION_SEARCH_START, m(XKB_MOD_NAME_CTRL "+" XKB_MOD_NAME_SHIFT), {{XKB_KEY_r}}},
+        {BIND_ACTION_FONT_SIZE_UP, m(XKB_MOD_NAME_CTRL), {{XKB_KEY_plus}}},
+        {BIND_ACTION_FONT_SIZE_UP, m(XKB_MOD_NAME_CTRL), {{XKB_KEY_equal}}},
+        {BIND_ACTION_FONT_SIZE_UP, m(XKB_MOD_NAME_CTRL), {{XKB_KEY_KP_Add}}},
+        {BIND_ACTION_FONT_SIZE_DOWN, m(XKB_MOD_NAME_CTRL), {{XKB_KEY_minus}}},
+        {BIND_ACTION_FONT_SIZE_DOWN, m(XKB_MOD_NAME_CTRL), {{XKB_KEY_KP_Subtract}}},
+        {BIND_ACTION_FONT_SIZE_RESET, m(XKB_MOD_NAME_CTRL), {{XKB_KEY_0}}},
+        {BIND_ACTION_FONT_SIZE_RESET, m(XKB_MOD_NAME_CTRL), {{XKB_KEY_KP_0}}},
+        {BIND_ACTION_SPAWN_TERMINAL, m(XKB_MOD_NAME_CTRL "+" XKB_MOD_NAME_SHIFT), {{XKB_KEY_n}}},
+        {BIND_ACTION_SHOW_URLS_LAUNCH, m(XKB_MOD_NAME_CTRL "+" XKB_MOD_NAME_SHIFT), {{XKB_KEY_o}}},
+        {BIND_ACTION_UNICODE_INPUT, m(XKB_MOD_NAME_CTRL "+" XKB_MOD_NAME_SHIFT), {{XKB_KEY_u}}},
+        {BIND_ACTION_PROMPT_PREV, m(XKB_MOD_NAME_CTRL "+" XKB_MOD_NAME_SHIFT), {{XKB_KEY_z}}},
+        {BIND_ACTION_PROMPT_NEXT, m(XKB_MOD_NAME_CTRL "+" XKB_MOD_NAME_SHIFT), {{XKB_KEY_x}}},
     };
 
     conf->bindings.key.count = ALEN(bindings);
@@ -2872,46 +2909,47 @@ add_default_key_bindings(struct config *conf)
 static void
 add_default_search_bindings(struct config *conf)
 {
-    static const struct config_key_binding bindings[] = {
-        {BIND_ACTION_SEARCH_SCROLLBACK_UP_PAGE, m_shift, {{XKB_KEY_Prior}}},
-        {BIND_ACTION_SEARCH_SCROLLBACK_DOWN_PAGE, m_shift, {{XKB_KEY_Next}}},
-        {BIND_ACTION_SEARCH_CANCEL, m_ctrl, {{XKB_KEY_c}}},
-        {BIND_ACTION_SEARCH_CANCEL, m_ctrl, {{XKB_KEY_g}}},
-        {BIND_ACTION_SEARCH_CANCEL, m_none, {{XKB_KEY_Escape}}},
-        {BIND_ACTION_SEARCH_COMMIT, m_none, {{XKB_KEY_Return}}},
-        {BIND_ACTION_SEARCH_FIND_PREV, m_ctrl, {{XKB_KEY_r}}},
-        {BIND_ACTION_SEARCH_FIND_NEXT, m_ctrl, {{XKB_KEY_s}}},
-        {BIND_ACTION_SEARCH_EDIT_LEFT, m_none, {{XKB_KEY_Left}}},
-        {BIND_ACTION_SEARCH_EDIT_LEFT, m_ctrl, {{XKB_KEY_b}}},
-        {BIND_ACTION_SEARCH_EDIT_LEFT_WORD, m_ctrl, {{XKB_KEY_Left}}},
-        {BIND_ACTION_SEARCH_EDIT_LEFT_WORD, m_alt, {{XKB_KEY_b}}},
-        {BIND_ACTION_SEARCH_EDIT_RIGHT, m_none, {{XKB_KEY_Right}}},
-        {BIND_ACTION_SEARCH_EDIT_RIGHT, m_ctrl, {{XKB_KEY_f}}},
-        {BIND_ACTION_SEARCH_EDIT_RIGHT_WORD, m_ctrl, {{XKB_KEY_Right}}},
-        {BIND_ACTION_SEARCH_EDIT_RIGHT_WORD, m_alt, {{XKB_KEY_f}}},
-        {BIND_ACTION_SEARCH_EDIT_HOME, m_none, {{XKB_KEY_Home}}},
-        {BIND_ACTION_SEARCH_EDIT_HOME, m_ctrl, {{XKB_KEY_a}}},
-        {BIND_ACTION_SEARCH_EDIT_END, m_none, {{XKB_KEY_End}}},
-        {BIND_ACTION_SEARCH_EDIT_END, m_ctrl, {{XKB_KEY_e}}},
-        {BIND_ACTION_SEARCH_DELETE_PREV, m_none, {{XKB_KEY_BackSpace}}},
-        {BIND_ACTION_SEARCH_DELETE_PREV_WORD, m_ctrl, {{XKB_KEY_BackSpace}}},
-        {BIND_ACTION_SEARCH_DELETE_PREV_WORD, m_alt, {{XKB_KEY_BackSpace}}},
-        {BIND_ACTION_SEARCH_DELETE_NEXT, m_none, {{XKB_KEY_Delete}}},
-        {BIND_ACTION_SEARCH_DELETE_NEXT_WORD, m_ctrl, {{XKB_KEY_Delete}}},
-        {BIND_ACTION_SEARCH_DELETE_NEXT_WORD, m_alt, {{XKB_KEY_d}}},
-        {BIND_ACTION_SEARCH_EXTEND_CHAR, m_shift, {{XKB_KEY_Right}}},
-        {BIND_ACTION_SEARCH_EXTEND_WORD, m_ctrl, {{XKB_KEY_w}}},
-        {BIND_ACTION_SEARCH_EXTEND_WORD, m_ctrl_shift, {{XKB_KEY_Right}}},
-        {BIND_ACTION_SEARCH_EXTEND_WORD_WS, m_ctrl_shift, {{XKB_KEY_w}}},
-        {BIND_ACTION_SEARCH_EXTEND_LINE_DOWN, m_shift, {{XKB_KEY_Down}}},
-        {BIND_ACTION_SEARCH_EXTEND_BACKWARD_CHAR, m_shift, {{XKB_KEY_Left}}},
-        {BIND_ACTION_SEARCH_EXTEND_BACKWARD_WORD, m_ctrl_shift, {{XKB_KEY_Left}}},
-        {BIND_ACTION_SEARCH_EXTEND_LINE_UP, m_shift, {{XKB_KEY_Up}}},
-        {BIND_ACTION_SEARCH_CLIPBOARD_PASTE, m_ctrl, {{XKB_KEY_v}}},
-        {BIND_ACTION_SEARCH_CLIPBOARD_PASTE, m_ctrl_shift, {{XKB_KEY_v}}},
-        {BIND_ACTION_SEARCH_CLIPBOARD_PASTE, m_ctrl, {{XKB_KEY_y}}},
-        {BIND_ACTION_SEARCH_CLIPBOARD_PASTE, m_none, {{XKB_KEY_XF86Paste}}},
-        {BIND_ACTION_SEARCH_PRIMARY_PASTE, m_shift, {{XKB_KEY_Insert}}},
+    const struct config_key_binding bindings[] = {
+        {BIND_ACTION_SEARCH_SCROLLBACK_UP_PAGE, m(XKB_MOD_NAME_SHIFT), {{XKB_KEY_Prior}}},
+        {BIND_ACTION_SEARCH_SCROLLBACK_DOWN_PAGE, m(XKB_MOD_NAME_SHIFT), {{XKB_KEY_Next}}},
+        {BIND_ACTION_SEARCH_CANCEL, m(XKB_MOD_NAME_CTRL), {{XKB_KEY_c}}},
+        {BIND_ACTION_SEARCH_CANCEL, m(XKB_MOD_NAME_CTRL), {{XKB_KEY_g}}},
+        {BIND_ACTION_SEARCH_CANCEL, m("none"), {{XKB_KEY_Escape}}},
+        {BIND_ACTION_SEARCH_COMMIT, m("none"), {{XKB_KEY_Return}}},
+        {BIND_ACTION_SEARCH_FIND_PREV, m(XKB_MOD_NAME_CTRL), {{XKB_KEY_r}}},
+        {BIND_ACTION_SEARCH_FIND_NEXT, m(XKB_MOD_NAME_CTRL), {{XKB_KEY_s}}},
+        {BIND_ACTION_SEARCH_EDIT_LEFT, m("none"), {{XKB_KEY_Left}}},
+        {BIND_ACTION_SEARCH_EDIT_LEFT, m(XKB_MOD_NAME_CTRL), {{XKB_KEY_b}}},
+        {BIND_ACTION_SEARCH_EDIT_LEFT_WORD, m(XKB_MOD_NAME_CTRL), {{XKB_KEY_Left}}},
+        {BIND_ACTION_SEARCH_EDIT_LEFT_WORD, m(XKB_MOD_NAME_ALT), {{XKB_KEY_b}}},
+        {BIND_ACTION_SEARCH_EDIT_RIGHT, m("none"), {{XKB_KEY_Right}}},
+        {BIND_ACTION_SEARCH_EDIT_RIGHT, m(XKB_MOD_NAME_CTRL), {{XKB_KEY_f}}},
+        {BIND_ACTION_SEARCH_EDIT_RIGHT_WORD, m(XKB_MOD_NAME_CTRL), {{XKB_KEY_Right}}},
+        {BIND_ACTION_SEARCH_EDIT_RIGHT_WORD, m(XKB_MOD_NAME_ALT), {{XKB_KEY_f}}},
+        {BIND_ACTION_SEARCH_EDIT_HOME, m("none"), {{XKB_KEY_Home}}},
+        {BIND_ACTION_SEARCH_EDIT_HOME, m(XKB_MOD_NAME_CTRL), {{XKB_KEY_a}}},
+        {BIND_ACTION_SEARCH_EDIT_END, m("none"), {{XKB_KEY_End}}},
+        {BIND_ACTION_SEARCH_EDIT_END, m(XKB_MOD_NAME_CTRL), {{XKB_KEY_e}}},
+        {BIND_ACTION_SEARCH_DELETE_PREV, m("none"), {{XKB_KEY_BackSpace}}},
+        {BIND_ACTION_SEARCH_DELETE_PREV_WORD, m(XKB_MOD_NAME_CTRL), {{XKB_KEY_BackSpace}}},
+        {BIND_ACTION_SEARCH_DELETE_PREV_WORD, m(XKB_MOD_NAME_ALT), {{XKB_KEY_BackSpace}}},
+        {BIND_ACTION_SEARCH_DELETE_NEXT, m("none"), {{XKB_KEY_Delete}}},
+        {BIND_ACTION_SEARCH_DELETE_NEXT_WORD, m(XKB_MOD_NAME_CTRL), {{XKB_KEY_Delete}}},
+        {BIND_ACTION_SEARCH_DELETE_NEXT_WORD, m(XKB_MOD_NAME_ALT), {{XKB_KEY_d}}},
+        {BIND_ACTION_SEARCH_EXTEND_CHAR, m(XKB_MOD_NAME_SHIFT), {{XKB_KEY_Right}}},
+        {BIND_ACTION_SEARCH_EXTEND_WORD, m(XKB_MOD_NAME_CTRL), {{XKB_KEY_w}}},
+        {BIND_ACTION_SEARCH_EXTEND_WORD, m(XKB_MOD_NAME_CTRL "+" XKB_MOD_NAME_SHIFT), {{XKB_KEY_Right}}},
+        {BIND_ACTION_SEARCH_EXTEND_WORD, m(XKB_MOD_NAME_CTRL), {{XKB_KEY_w}}},
+        {BIND_ACTION_SEARCH_EXTEND_WORD_WS, m(XKB_MOD_NAME_CTRL "+" XKB_MOD_NAME_SHIFT), {{XKB_KEY_w}}},
+        {BIND_ACTION_SEARCH_EXTEND_LINE_DOWN, m(XKB_MOD_NAME_SHIFT), {{XKB_KEY_Down}}},
+        {BIND_ACTION_SEARCH_EXTEND_BACKWARD_CHAR, m(XKB_MOD_NAME_SHIFT), {{XKB_KEY_Left}}},
+        {BIND_ACTION_SEARCH_EXTEND_BACKWARD_WORD, m(XKB_MOD_NAME_CTRL "+" XKB_MOD_NAME_SHIFT), {{XKB_KEY_Left}}},
+        {BIND_ACTION_SEARCH_EXTEND_LINE_UP, m(XKB_MOD_NAME_SHIFT), {{XKB_KEY_Up}}},
+        {BIND_ACTION_SEARCH_CLIPBOARD_PASTE, m(XKB_MOD_NAME_CTRL), {{XKB_KEY_v}}},
+        {BIND_ACTION_SEARCH_CLIPBOARD_PASTE, m(XKB_MOD_NAME_CTRL "+" XKB_MOD_NAME_SHIFT), {{XKB_KEY_v}}},
+        {BIND_ACTION_SEARCH_CLIPBOARD_PASTE, m(XKB_MOD_NAME_CTRL), {{XKB_KEY_y}}},
+        {BIND_ACTION_SEARCH_CLIPBOARD_PASTE, m("none"), {{XKB_KEY_XF86Paste}}},
+        {BIND_ACTION_SEARCH_PRIMARY_PASTE, m(XKB_MOD_NAME_SHIFT), {{XKB_KEY_Insert}}},
     };
 
     conf->bindings.search.count = ALEN(bindings);
@@ -2922,12 +2960,12 @@ add_default_search_bindings(struct config *conf)
 static void
 add_default_url_bindings(struct config *conf)
 {
-    static const struct config_key_binding bindings[] = {
-        {BIND_ACTION_URL_CANCEL, m_ctrl, {{XKB_KEY_c}}},
-        {BIND_ACTION_URL_CANCEL, m_ctrl, {{XKB_KEY_g}}},
-        {BIND_ACTION_URL_CANCEL, m_ctrl, {{XKB_KEY_d}}},
-        {BIND_ACTION_URL_CANCEL, m_none, {{XKB_KEY_Escape}}},
-        {BIND_ACTION_URL_TOGGLE_URL_ON_JUMP_LABEL, m_none, {{XKB_KEY_t}}},
+    const struct config_key_binding bindings[] = {
+        {BIND_ACTION_URL_CANCEL, m(XKB_MOD_NAME_CTRL), {{XKB_KEY_c}}},
+        {BIND_ACTION_URL_CANCEL, m(XKB_MOD_NAME_CTRL), {{XKB_KEY_g}}},
+        {BIND_ACTION_URL_CANCEL, m(XKB_MOD_NAME_CTRL), {{XKB_KEY_d}}},
+        {BIND_ACTION_URL_CANCEL, m("none"), {{XKB_KEY_Escape}}},
+        {BIND_ACTION_URL_TOGGLE_URL_ON_JUMP_LABEL, m("none"), {{XKB_KEY_t}}},
     };
 
     conf->bindings.url.count = ALEN(bindings);
@@ -2938,18 +2976,18 @@ add_default_url_bindings(struct config *conf)
 static void
 add_default_mouse_bindings(struct config *conf)
 {
-    static const struct config_key_binding bindings[] = {
-        {BIND_ACTION_SCROLLBACK_UP_MOUSE, m_none, {.m = {BTN_BACK, 1}}},
-        {BIND_ACTION_SCROLLBACK_DOWN_MOUSE, m_none, {.m = {BTN_FORWARD, 1}}},
-        {BIND_ACTION_PRIMARY_PASTE, m_none, {.m = {BTN_MIDDLE, 1}}},
-        {BIND_ACTION_SELECT_BEGIN, m_none, {.m = {BTN_LEFT, 1}}},
-        {BIND_ACTION_SELECT_BEGIN_BLOCK, m_ctrl, {.m = {BTN_LEFT, 1}}},
-        {BIND_ACTION_SELECT_EXTEND, m_none, {.m = {BTN_RIGHT, 1}}},
-        {BIND_ACTION_SELECT_EXTEND_CHAR_WISE, m_ctrl, {.m = {BTN_RIGHT, 1}}},
-        {BIND_ACTION_SELECT_WORD, m_none, {.m = {BTN_LEFT, 2}}},
-        {BIND_ACTION_SELECT_WORD_WS, m_ctrl, {.m = {BTN_LEFT, 2}}},
-        {BIND_ACTION_SELECT_QUOTE, m_none, {.m = {BTN_LEFT, 3}}},
-        {BIND_ACTION_SELECT_ROW, m_none, {.m = {BTN_LEFT, 4}}},
+    const struct config_key_binding bindings[] = {
+        {BIND_ACTION_SCROLLBACK_UP_MOUSE, m("none"), {.m = {BTN_BACK, 1}}},
+        {BIND_ACTION_SCROLLBACK_DOWN_MOUSE, m("none"), {.m = {BTN_FORWARD, 1}}},
+        {BIND_ACTION_PRIMARY_PASTE, m("none"), {.m = {BTN_MIDDLE, 1}}},
+        {BIND_ACTION_SELECT_BEGIN, m("none"), {.m = {BTN_LEFT, 1}}},
+        {BIND_ACTION_SELECT_BEGIN_BLOCK, m(XKB_MOD_NAME_CTRL), {.m = {BTN_LEFT, 1}}},
+        {BIND_ACTION_SELECT_EXTEND, m("none"), {.m = {BTN_RIGHT, 1}}},
+        {BIND_ACTION_SELECT_EXTEND_CHAR_WISE, m(XKB_MOD_NAME_CTRL), {.m = {BTN_RIGHT, 1}}},
+        {BIND_ACTION_SELECT_WORD, m("none"), {.m = {BTN_LEFT, 2}}},
+        {BIND_ACTION_SELECT_WORD_WS, m(XKB_MOD_NAME_CTRL), {.m = {BTN_LEFT, 2}}},
+        {BIND_ACTION_SELECT_QUOTE, m("none"), {.m = {BTN_LEFT, 3}}},
+        {BIND_ACTION_SELECT_ROW, m("none"), {.m = {BTN_LEFT, 4}}},
     };
 
     conf->bindings.mouse.count = ALEN(bindings);
@@ -3064,12 +3102,7 @@ config_load(struct config *conf, const char *conf_path,
         .mouse = {
             .hide_when_typing = false,
             .alternate_scroll_mode = true,
-            .selection_override_modifiers = {
-                .shift = true,
-                .alt = false,
-                .ctrl = false,
-                .super = false,
-            },
+            .selection_override_modifiers = tll_init(),
         },
         .csd = {
             .preferred = CONF_CSD_PREFER_SERVER,
@@ -3125,6 +3158,7 @@ config_load(struct config *conf, const char *conf_path,
     };
 
     memcpy(conf->colors.table, default_color_table, sizeof(default_color_table));
+    parse_modifiers(XKB_MOD_NAME_SHIFT, 5, &conf->mouse.selection_override_modifiers);
 
     tokenize_cmdline("notify-send -a ${app-id} -i ${app-id} ${title} ${body}",
                      &conf->notify.argv.args);
@@ -3324,6 +3358,9 @@ key_binding_list_clone(struct config_key_binding_list *dst,
         struct config_key_binding *new = &dst->arr[i];
 
         *new = *old;
+        memset(&new->modifiers, 0, sizeof(new->modifiers));
+        tll_foreach(old->modifiers, it)
+            tll_push_back(new->modifiers, xstrdup(it->item));
 
         switch (old->aux.type) {
         case BINDING_AUX_NONE:
@@ -3399,6 +3436,11 @@ config_clone(const struct config *old)
 
     conf->env_vars.length = 0;
     conf->env_vars.head = conf->env_vars.tail = NULL;
+
+    memset(&conf->mouse.selection_override_modifiers, 0, sizeof(conf->mouse.selection_override_modifiers));
+    tll_foreach(old->mouse.selection_override_modifiers, it)
+        tll_push_back(conf->mouse.selection_override_modifiers, xstrdup(it->item));
+
     tll_foreach(old->env_vars, it) {
         struct env_var copy = {
             .name = xstrdup(it->item.name),
@@ -3431,13 +3473,13 @@ UNITTEST
     bool ret = config_load(&original, "/dev/null", &nots, &overrides, false, false);
     xassert(ret);
 
-    struct config *clone = config_clone(&original);
-    xassert(clone != NULL);
-    xassert(clone != &original);
+    //struct config *clone = config_clone(&original);
+    //xassert(clone != NULL);
+    //xassert(clone != &original);
 
     config_free(&original);
-    config_free(clone);
-    free(clone);
+    //config_free(clone);
+    //free(clone);
 
     fcft_fini();
 
@@ -3473,6 +3515,7 @@ config_free(struct config *conf)
     free_key_binding_list(&conf->bindings.search);
     free_key_binding_list(&conf->bindings.url);
     free_key_binding_list(&conf->bindings.mouse);
+    tll_free_and_free(conf->mouse.selection_override_modifiers, free);
 
     tll_foreach(conf->env_vars, it) {
         free(it->item.name);
@@ -3603,6 +3646,7 @@ check_if_font_is_monospaced(const char *pattern,
     return is_monospaced;
 }
 
+#if 0
 xkb_mod_mask_t
 conf_modifiers_to_mask(const struct seat *seat,
                        const struct config_key_modifiers *modifiers)
@@ -3618,3 +3662,4 @@ conf_modifiers_to_mask(const struct seat *seat,
         mods |= modifiers->super << seat->kbd.mod_super;
     return mods;
 }
+#endif