+++ config.h
@@ -38,12 +38,14 @@ struct config_font {
 };
 DEFINE_LIST(struct config_font);
 
+#if 0
 struct config_key_modifiers {
     bool shift;
     bool alt;
     bool ctrl;
     bool super;
 };
+#endif
 
 struct argv {
     char **args;
@@ -74,9 +76,12 @@ enum key_binding_type {
     MOUSE_BINDING,
 };
 
+typedef tll(char *) config_modifier_list_t;
+
 struct config_key_binding {
     int action;  /* One of the varios bind_action_* enums from wayland.h */
-    struct config_key_modifiers modifiers;
+    //struct config_key_modifiers modifiers;
+    config_modifier_list_t modifiers;
     union {
         /* Key bindings */
         struct {
@@ -263,7 +268,8 @@ struct config {
     struct {
         bool hide_when_typing;
         bool alternate_scroll_mode;
-        struct config_key_modifiers selection_override_modifiers;
+        //struct config_key_modifiers selection_override_modifiers;
+        config_modifier_list_t selection_override_modifiers;
     } mouse;
 
     struct {
@@ -375,10 +381,11 @@ struct config *config_clone(const struct config *old);
 bool config_font_parse(const char *pattern, struct config_font *font);
 void config_font_list_destroy(struct config_font_list *font_list);
 
+#if 0
 struct seat;
 xkb_mod_mask_t
 conf_modifiers_to_mask(
     const struct seat *seat, const struct config_key_modifiers *modifiers);
-
+#endif
 bool check_if_font_is_monospaced(
     const char *pattern, user_notifications_t *notifications);