+++ CHANGELOG.md
@@ -56,6 +56,10 @@
   numlock is ignored.
 * A new `resize-by-cells` option, enabled by default, allows the size
   of floating windows to be constrained to multiples of the cell size.
+* Support for custom (i.e. other than ctrl/shift/alt/super) modifiers
+  in key bindings ([#1348][1348]).
+
+[1348]: https://codeberg.org/dnkl/foot/issues/1348
 
 
 ### Changed